requirements.txt
python-binance==1.0.19
websockets==12.0
xgboost==2.0.3
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
scikit-learn==1.3.2
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0